Move escape and unescape to names.h (NFC)
This moves `escape` from `WasmBinaryReader` and `unescape` from
`wasm-split.cpp` to `names.h`.

Currently `unescape` is a helper function only used in `wasm-split.cpp`,
and this is a preparation for wasm-merge to use `unescape`. This also
moves `escape`, which is a pair, from `WasmBinaryReader`, because it is
not necessarily related to binary anyway.
