+++ src/wasm-binary.h
@@ -1780,7 +1780,6 @@ public:
 
   void readTags();
 
-  static Name escape(Name name);
   void readNames(size_t sectionPos, size_t payloadLen);
   void readFeatures(size_t sectionPos, size_t payloadLen);
   void readDylink(size_t payloadLen);
